Task 2: Use JPMorgan Chase & Co. frameworks and tools
Purpose: The objective of this task will be for you to fix the client-side web application so that it displays a graph that automatically updates as it gets data from the server application. Currently, the web application only gets data every time you click on the 'Start Streaming Data' button and does not aggregate duplicated data.