Task 3: Display data visually for traders
Purpose : You will use perspective to generate a live graph that displays the data feed in a clear and visually appealing way for traders to monitor. Recall that the purpose of this graph is to monitor and determine when a trading opportunity may arise as a result of the temporary weakening of a correlation between two stock prices. Given this graph, the trader should be able to quickly and easily notice when the ratio moves too far from the average historical correlation. In the first instance, we'll assume that threshold is +/-10% of the 12-month historical average ratio.